+++ src/AssetsAutoCompressComponent.php
@@ -8,7 +8,6 @@
 
 namespace skeeks\yii2\assetsAuto;
 
-use skeeks\yii2\assetsAuto\vendor\HtmlCompressor;
 use yii\base\BootstrapInterface;
 use yii\base\Component;
 use yii\base\Event;
@@ -74,6 +73,12 @@ class AssetsAutoCompressComponent extends Component implements BootstrapInterfac
     public $cssFileCompile = true;
 
     /**
+     * Enables the compilation of files in groups rather than in a single file. Works only when the $cssFileCompile option is enabled
+     * @var bool 
+     */
+    public $cssFileCompileByGroups = false;
+
+    /**
      * Trying to get css files to which the specified path as the remote file, skchat him to her.
      * @var bool
      */
@@ -103,6 +108,12 @@ class AssetsAutoCompressComponent extends Component implements BootstrapInterfac
      * @var bool
      */
     public $jsFileCompile = true;
+
+    /**
+     * Enables the compilation of files in groups rather than in a single file. Works only when the $jsFileCompile option is enabled
+     * @var bool
+     */
+    public $jsFileCompileByGroups = false;
 
     /**
      * @var array
@@ -218,7 +229,7 @@ class AssetsAutoCompressComponent extends Component implements BootstrapInterfac
             $app->response->on(\yii\web\Response::EVENT_BEFORE_SEND, function (\yii\base\Event $event) use ($app) {
                 $response = $event->sender;
 
-                if ($this->enabled && ($this->htmlFormatter instanceof IFormatter)  && $response->format == \yii\web\Response::FORMAT_HTML && !$app->request->isAjax && !$app->request->isPjax) {
+                if ($this->enabled && ($this->htmlFormatter instanceof IFormatter) && $response->format == \yii\web\Response::FORMAT_HTML && !$app->request->isAjax && !$app->request->isPjax) {
                     if (!empty($response->data)) {
                         $response->data = $this->_processingHtml($response->data);
                     }
@@ -236,15 +247,21 @@ class AssetsAutoCompressComponent extends Component implements BootstrapInterfac
     protected function _processing(View $view)
     {
         //Компиляция файлов js в один.
+        //echo "<pre><code>" . print_r($view->jsFiles, true);die;
         if ($view->jsFiles && $this->jsFileCompile) {
             \Yii::beginProfile('Compress js files');
             foreach ($view->jsFiles as $pos => $files) {
                 if ($files) {
-                    $view->jsFiles[$pos] = $this->_processingJsFiles($files);
+                    if ($this->jsFileCompileByGroups) {
+                        $view->jsFiles[$pos] = $this->_processAndGroupJsFiles($files);
+                    } else {
+                        $view->jsFiles[$pos] = $this->_processingJsFiles($files);
+                    }
                 }
             }
             \Yii::endProfile('Compress js files');
         }
+        //echo "<pre><code>" . print_r($view->jsFiles, true);die;
 
         //Compiling js code that is found in the html code of the page.
         if ($view->js && $this->jsCompress) {
@@ -261,8 +278,11 @@ class AssetsAutoCompressComponent extends Component implements BootstrapInterfac
         //Compiling css files
         if ($view->cssFiles && $this->cssFileCompile) {
             \Yii::beginProfile('Compress css files');
-
-            $view->cssFiles = $this->_processingCssFiles($view->cssFiles);
+            if ($this->cssFileCompileByGroups) {
+                $view->cssFiles = $this->_processAndGroupCssFiles($view->cssFiles);
+            } else {
+                $view->cssFiles = $this->_processingCssFiles($view->cssFiles);
+            }
             \Yii::endProfile('Compress css files');
         }
 
@@ -313,6 +333,55 @@ JS
             \Yii::endProfile('Moving css files bottom');
         }
     }
+
+    /**
+     * @param array $files
+     */
+    protected function _processAndGroupJsFiles($files = [])
+    {
+        if (!$files) {
+            return [];
+        }
+
+        $result = [];
+        $groupedFiles = $this->_getGroupedFiles($files);
+        foreach ($groupedFiles as $files)
+        {
+            $resultGroup = $this->_processingJsFiles($files);
+            $result = ArrayHelper::merge($result, $resultGroup);
+        }
+        
+        return $result;
+        echo "<pre><code>" . print_r($result, true); die;
+
+    }
+
+    public function _getGroupedFiles($files)
+    {
+        $result = [];
+
+        $lastKey = null;
+        $tmpData = [];
+        $counter = 0;
+        foreach ($files as $fileCode => $fileTag) {
+            list($one, $two, $key) = explode("/", $fileCode);
+
+            $counter ++;
+
+            if ($key != $lastKey && $counter > 1) {
+                $result[] = $tmpData;
+                $tmpData = [];
+                $tmpData[$fileCode] = $fileTag;
+            } else {
+                $tmpData[$fileCode] = $fileTag;
+            }
+
+            $lastKey = $key;
+        }
+
+        return $result;
+    }
+
     /**
      * @param array $files
      * @return array
@@ -435,7 +504,7 @@ JS
             throw new \Exception("Unable to open file: '{$filePath}'");
         }
         $filesSize = filesize($filePath);
-        if($filesSize){
+        if ($filesSize) {
             return fread($file, $filesSize);
         }
         fclose($file);
@@ -481,6 +550,28 @@ JS
 
         return $result;
     }
+    
+    /**
+     * @param array $files
+     */
+    protected function _processAndGroupCssFiles($files = [])
+    {
+        if (!$files) {
+            return [];
+        }
+
+        $result = [];
+        $groupedFiles = $this->_getGroupedFiles($files);
+        foreach ($groupedFiles as $files)
+        {
+            $resultGroup = $this->_processingCssFiles($files);
+            $result = ArrayHelper::merge($result, $resultGroup);
+        }
+        
+        return $result;
+
+    }
+    
     /**
      * @param array $files
      * @return array
@@ -616,9 +707,9 @@ JS
     {
         if ($this->htmlFormatter instanceof IFormatter) {
             $r = new \ReflectionClass($this->htmlFormatter);
-            \Yii::beginProfile('Format html: ' . $r->getName());
-                $result = $this->htmlFormatter->format($html);
-            \Yii::endProfile('Format html: ' . $r->getName());
+            \Yii::beginProfile('Format html: '.$r->getName());
+            $result = $this->htmlFormatter->format($html);
+            \Yii::endProfile('Format html: '.$r->getName());
             return $result;
         }
 
@@ -629,9 +720,9 @@ JS
 
 
     /**
-     * @deprecated >= 1.4
      * @param $value
      * @return $this
+     * @deprecated >= 1.4
      */
     public function setHtmlCompress($value)
     {
@@ -639,20 +730,20 @@ JS
     }
 
     /**
-     * @deprecated >= 1.4
      * @param $value
      * @return $this
+     * @deprecated >= 1.4
      */
     public function getHtmlCompress()
     {
         return $this;
     }
     /**
-     * @deprecated >= 1.4
      * @param $value array options for compressing output result
      *   * extra - use more compact algorithm
      *   * no-comments - cut all the html comments
      * @return $this
+     * @deprecated >= 1.4
      */
     public function setHtmlCompressOptions($value)
     {
@@ -660,9 +751,9 @@ JS
     }
 
     /**
-     * @deprecated >= 1.4
      * @param $value
      * @return $this
+     * @deprecated >= 1.4
      */
     public function getHtmlCompressOptions()
     {